+++ RULEngine/controller.py
@@ -1,17 +1,15 @@
 # Under MIT licence, see LICENCE.txt
 
 import logging
+from collections import namedtuple
 from multiprocessing import Queue
 from queue import Empty
-
 from typing import Dict
-from collections import namedtuple
 
-from RULEngine.robot import Robot
 from RULEngine.controllers import VelocityController, PositionController
-
+from RULEngine.filters.path_smoother import path_smoother
+from RULEngine.robot import Robot
 from Util.constant import PLAYER_PER_TEAM
-from Util.path_smoother import path_smoother
 from config.config_service import ConfigService
 
 RobotPacket = namedtuple('RobotPacket', 'robot_id command kick_type kick_force dribbler_active charge_kick')
