+++ RULEngine/controllers/position_controllers.py
@@ -1,9 +1,8 @@
 
+from RULEngine.controllers.PID import PID
 from RULEngine.controllers.controller_base_class import ControllerBaseClass
 from RULEngine.robot import Robot
-
 from Util import Pose
-from Util.PID import PID
 from Util.geometry import wrap_to_pi, rotate
